(-2 [*]  4/4 unreliables.. ((.. 3752)) )
(many legacy systems were using the old deprecated style of
selective listener that would be allowed to determine the
result of the performer. this is now deprecated as a smell -
if a performer wants to imbue an argument proc with this power,
it must be through special arrangement; and not thru the
standard selective listener. we did not detect the severity
of this old technique until we changed our "expect event"
listener to result in a true-ish.)

  - the temporary _GREENLIST_ has reached a chokepoint so we
    have divided it into two (_GL_2_).

(2 in 19.79)