(   [tmx] map an attribute - two column output..)
this is a bad example because i’m not using `awk` as well as
i could be, but here is the general idea:

    tmx map "is lib" | grep ' lib$' | awk '{print $1}' | wc -l

for the first time ever, i can programmatically count how
many library sidesystems i have. yay. this idea can be extended
to other attributes and how each attribute is expressed is totally
up to arbitrary code.
.