(-5.4 slowie: integrate that with real tmx..)
(no coverage because the integration itself is a bit of an afterthought
for now, and we want the focus rather to be on covering operations
as API calls. likely CLI will get more complicated soon and we should
cover then.)

it’s not as pretty as we want it to be (and as it can be) but
here’s how it looks when our API operations are adapted directly
to CLI:

    tmx test-all -list-files -test-directory tmx/test -test-directory zerk/test

we have opened ourselves to the possibility of writing parameters to
the API invocation directly for some, hence the addition in API API
in slowie.

try to abstract some of the expression adaptation for CLI in [tmx]/
.