(-5.3 [*] first whack good enough for now help screens.. (( 72 )) )
Synopsis: with a coarse version of help screens working (of 1,3,5
  we have 1,3 and 2-ish displaying, known issues noted below) and the
  essential functionality of “counts” working, we can move on for now
  towards finally getting a usable	“test all” again..

Gripping narrative:

at the start of the commit we had the help screen renderer facilities
(one for branch and one for terminal). but we had to maintain two
different full calls based on what kind of (what we now call)
“operator” we were expressing (branch vs. terminal).

furthermore, different calls were necessary depending on the various
architectural conventions used by the “splays” of operations in
whatever service was being called against.

so that become one “design objective” was to corral these calls
down to one somehow.

the “design consequences” of this were multi-step: for one thing we
conceived of (and then formalized) this “didactics” structure for the
reasons justified there ([#ze-055], spliced into [#br-098]).

and another design consequence of this (and perhaps the most salient
one here) is this: we now have this idea of emitting an
`operator_resolved` emission. this allows the “modality client” to
maintain a “selection stack”, an idea we have relied upon heavily
in our forebears. this was A) necessary to express help screens in the
way we have architected (which we like, mostly) and B) it is an
improvement upon on the hacks we were doing to allow the outermost
client to reach the innermost operator for modality customizations
(reaching from CLI client to API client to API “current operation”).

Details design consequences:

  - so that branchy and not-branchy operators can share the same
    reflection methods with the same name, we have added one
    dimension of depth to our much revered uber-simple interface
    that used streams of plain old symbols. now, we must know if
    these items are primaries or “operations” (we might remove
    this distinction one day, but for now it is here). this
    categorical first term of the tuple then determines the
    section name of the group of items displayed in help screens.

KNOWN ISSUES:

while generated help screens *do* reflect hacks done by the
argument scanner, they do *not* reflect hacks done by the “feature
injection” thing. saved for later because a priority is to get
“test all” working

q zerk/test/5-API/42-argument-scanner_spec.rb zerk/test/4-argument-scanner zerk/test/6-non-interactive-CLI/42-multi-mode-argument-scanner test_support/test/48-slowie/54-operations brazen/test/cli-support/when/help/screen-for-branch_spec.rb tmx/test/64-cli -order