(-4.16 thru -4.14 [tmx]=>([co],[st],[ts],gi]) ..)
(still not integrating with the universe - only the tests that were
moved in this commit are green. integrating with the universe at the
end, once we have a better sense for what it will take to get
description lines and usage structures (anew for this new tmx client).)

in this commit we restore what we now call [#br-083] - the thing where
binary executable one-off’s were loaded assuming certain conventions.
the implementing code now use the the convention for the “standard 5”
arguments for a CLI. (many clients not covered here probably remain
to be changed)

  - [gi] “citxt” one-off attempted to feign modularity by making a
    one-off CLI base class and then subclassing it to implement the
    one-off client. because we now eschew using inheritance in this
    manner (preferring composition), we have undone this added
    complexity.
.