(     [ba] first splash at this linked list thing.. (( 12 )) )
Justification thru Narrative: the "multi-mode argument scanner"
started out being written in a cautious, would-be immutable
("shell"/"kernel" pattern) way. but then as it entered adolescence,
we introduced exposures to it to make it mutable (in certain ways)
at run time. this, then, obviated all the shell/kernel walls we
had so carefully put up around it, adding to the general noise of
the ever-growing file/node.

as the code this work will eventually replace suggests, (and although
there is literature touting the benefits of growable arrays over
linked lists), mutating one of our "polymorphic scanners" (with an
array as its internal datastore) is kludgy and ugly.

so we took a step back and asked, "what are the interations (reads
and writes) we want to have with our scanner list, and what data
structure(s) are best poised to implement such a thing?" and for
whatever reason we had this itch to scratch about linked lists.

we are trying to encapsulate the linked-listed-ness of the thing
inside its interface. hopefully this will allow us to flip the
implementation if we ever wanted to for some weird reason.

(it's worth mentioning that that the typical "multi-mode argument
scanner" gets only two or three scanner items pushed on to it
(and so two or three shifts at runtime), so this is really just an
excercize in aesthetics and other psychological neuroses.)

For fun:

just for fun, here are the timestamps from our commits that were
squashed into this commit. each one of these corresponds to when
we make a commit after passing a test (in the "three laws" way).

it's worth mentioning we had a rough version of this working (and
failing) in its target place before starting over here. also the
two time jumps below that are more than an hour, those included
physical transit and/or eating ice cream.

    20:08:30
    20:14:53 (5)
    20:15:58 (1)
    21:17:14 (61)
    21:19:10 (2)
    21:22:18 (3)
    22:38:45 (76)
    23:14:07 (25)
    23:35:41 (21)
    23:55:11 (20)
    23:58:51 (4)
.