(-3.8.D [tmx] 2/2 "snap" one-off complete..)
the snap standalone utility script is justified in the comment in
its file.

(as we anticipated, getting that IPC stuff “right” was tricky.
we are trying to ween off of using `popen3` and trying to understand
the underlying pipes a bit more..)

as we did before, for fun here's timetracking (which tries to capture
the whole start to finish of this script from first line of code to
its bleeding alpha version). instead of commiting what will be the
next commit using our `snap` basic alias (explained in the comment
committed here), we wanted the maiden voyage of this utility script
to be for that. so:

    01:18        - finished passing the last tests of our next commit
    03:26 (128)  - commited the basic "porcelain" of the CLI
    04:44  (78)  - started reviewing about the process stuff
    05:37  (53)  - now

around 4.5 hours ish. meh that's fine

  - [ts] (minuscule misspelling and extra space)
.