(-3.8   [ze] clean up how [ze] niCLI argument scanner is built (.. 3909))
Amusing anecdote:

the multi-mode argument scanner started life being architected around
a “shell”/“kernel” builder pattern because of the intricate
preparations that had to be made to distribute the right parts to
the right scanners to support all of its hacks.

however as time progressed we began to experiment with mutating its
“definitional” characteristics at runtime, something that emerged
as a requirement only later.

given that there were now runtime mutators for the front scanner,
having the clunky DSL class and “builder” class there too seemed
superfluous. the objective was rid the stack of these classes.

part of this grand dive into mutability involved using something like
a linked list for the scanners, rather that a polymorphic scanner
(which at its core was an array). so that was that last interlude
just now.

finally, armed with our new “ordered collection” class and ready
to rid the subject of the superfluity, we got pretty far along with
this approach util we hit a snafu: it turns out we *do* need the
shell/kernel builder pattern, because for example sub-scanners
must be built with listeners and typically in the definition block,
listeners aren’t set until the end of the definition.

but hopefully, now that we have rewritten almost the same thing from
scratch twice, it has gelled and cleaned up a little more.

  12:58       woke to begin
  19:45 (405) now

  - [bs] always cleaning up this name conventions file

(2 in 19.72)