(-3.7   [ze] break hash-based parse method into a parser builder (.. 3909))
if we are parsing multiple primaries off an argument scanner against a
single operator branch (imagine a hash) (this is how we typically
implement operations), we should not have to parse the request array
one time for each primary that we parse (because it is the same request
array each time).

here have broken the process up into dedicated performers:

  request array -> request -> matcher -> search -> item

now it will be possible to use the same matcher to produce
a series of items without having to recalculate the first
two arrows at each parse.

  19:47        started modularlize
  01:44 (357)  now

(2 in 20.17, 20.22)