((       [tab] introduce gem. refactor legacy minimal. (( 1 )) ))
This rally: the three or so commits before this one *at writing* need
to be squashed *somehow* so they clump into more uniformly sized pieces.
the problem is, we don’t know which commits to squash where.

because we want a simple visual report of the sizes of
a list of arbitrary commits (in terms of line numbers of the diffs,
or whatever), the time has come to make our table librar*ies* abstract
and modular enough to be able to handle stream of arbitrary data from
the shell. (like basically doing a tiny bit more than `awk` can do
with its form of table rendering.)

what we want is that with a shell one-off we produce a stream of
two-column data and pipe it into our new script. that script should
render a table with our “lipstick” visualization and whatever other
statistical summary data.

in the remainder of this rally we will attempt to unify *all* old
table libraries with a full overhaul using our new architecture.

This commit: the corefile of the incumbent (and now sunsetting) table
library (under [br]) had incidentally become a #feature-island. we
have taken its minimal but typical test use case as a way to drive
forward the foundation for this new architecture we want to see..

  (times are not from a contiguous session)
  18:36        create 'tabular' directory
  21:52 (2h+)  start in earnest researching
  22:06   (8)  start in earnest building up first test
  01:15 (3h+)  replaced original simplest of things, rough pass (needs refactor)
  01:47  (32)  good enough for now
.