((       [tab],[ze] mid-life cleanup.. (( 45 )) ))
Originally this commit started out as work towards supporting
a “sprintf” modifier. this lead to cleanup in the one file near type
mavens. but eventually we discovered a deeper issue that needed
cleanup:

when trying to add a feature, found names confusing along one particular
axis: because “summary fields” change the “position system” and because
the position system is a fundamental underpinning to so many parts of
the system, we wanted to make the names more clear as to exactly what
they are, giving a nod to the highly managed mutability of the
structures in the pipeline.

we SRP’d some things and un-SRP’d other things - in the latter category
we used to have a dup-and-mutate performer that effected one instance
per row (“tuple”) of input data. we found a way to accomplish the
logical objectives while using the same performer to manage the traversal
of all the input.

created “field survey writer” to bundle the many methods we were
exposing from that one client.

also we changed how defined fields are stored internally - now they
are in one flat array that isomorphs to the final rendered table,
rather than the arrangement we had before where summary fields were
kept separately.

corrected the name “field offset” to “input tuple offset” as
appropriate, with the newfound knowledge of [#050.B]. likewise
give other methods more descriptive names that acknowledge whether
or not the data is post-hooks and post-summary fields:

  - “number of all fields” not “number of fields”
  - “all defined fields” not “fields”
  - “every survey of every field”, not “field surveys”

q -order tabular/test zerk/test/5.5-CLI