((-6     [cm] ASCII matrix via hard-coded shapes-layers (( +8 )) ))
Discussion: hoping that that’s the hardest part. the ASCII rendering
component of this is just a stand-in for something more appropriate,
but as long as it’s usable it makes sense to target this as an output
modality; because it can be useful and there is some novelty to it.
however, it’s hard not to feel silly about sweating over some problems
that were long solved 30+ years ago.

  - [ts] sneak in this nicely advanced line expecter.

  06:37   first test (load the magnet)
  09:17   everything but fill rect. have to stop
  17:11   rebegan
  24:ish  slept
   5:ish  woke
  11:50   done

q code_metrics/test/44-magnetics/66-ascii-matrix-via-shapes-layers code_metrics/test/64-cli/92-mondrian_spec.rb code_metrics/test/64-cli/92-mondrian_spec.rb
-