(-3.293 [*] visual interlude: universe help (.. 4202))
during the planning phase we thought it would be more prudent to work
on the more basic issues of mounting and streaming the remote operators
first (just in terms of the low-level mechanics involved there) before
the “view”-like concern of help screens. however it turns out what
“feels” natural is to use the help screens to drive the development
of the mounting, etc..

this lead to the desire for an “extended” form of the help screen
that we can turn on/off with a -verbose switch. *this* in turn lead
us to develop more sophisticated parsing exposures in “no-deps”, which
we covered.

  - [tmx] we had made a load ticket class for sidesystems *again*,
    then merged the new work there in with the old load ticket

  - [ba] the superior “tailerer via separator” developed in [cm] has
    floated up to here.

  00:27  start to think about streaming the sidesystems for help
  07:22  begin verbose for help
  15:20  begining omnibus of tests for [ze] for new parse functions
  19:08  two levels of verbose look good
  15:48  now

(2 in 31.27, 29.74)