(-3.276 [ba]<-[*] cleanup of state machine (.. 4213))
Main cleanup:

(1) parsing logic was almost impossible to follow because of the
use of inheritance. at one time there was an “active” parsing
algorithm and a “passive” one. (the latter somehow pulled parsed
units on-demand, whereas the former works in the familiar way
by parsing the input all at once.) because we didn’t know better
at the time, we mistakenly thought it was prudent to create a class
hierarchy of two child parsers sharing one base set of methods. no,
wrong. the righter thing to do is to break the problem up into smaller,
isolated components.

also it was cost without benefit because we weren’t using passive
parsing anywhere anyway. (but we are about to next)

(2) improve the expressive power of handlers and future-proof them
somewhat by having the state handlers receive “state machine” instances
(actually a custom proxy) instead of just the user “downstream”. (the
user downstream is now accessed as a member of this “exposures proxy”.)

it should not be that there are special state names with magical
meaning. (for now we’ll allow `beginning` as an exception.) the way
`ending` was handled was kind of nasty. now there is the idea of
“directives” for this, which are just special methods you call on
the exposures proxy. at writing the only directive is `end` (or
similar) but more is up next.

Other cleanup:

dissolved nodes that were redundant with other nodes, without a clear
purpose. (this “grammar” class as if to be an immutable definition
structure, when what was not clear from the existing graph was that
the “state machine” class itself was exactly that.)

experiments that weren’t actually used anywhere. (this passive
parse thing, which we’re about to get into..)

overly-complicated parsing loop simplified. (we matched a state
with a “barrier to entry” proc by memoizing its resulted matchdata.
we used to then change the step method so that at the next step
there would be a dedicated step to processing this matchdata
from the previous step. if you process the matchdata right away
you never need to change your step method, and nothing was otherwise
gained from doing it this way.)

(2 in 20.17)