(-3.262 [tmx] deep help (mounted one-offs) (.. 4237))
this is a “sneak preview” of the bulk of the work for this without
*any* coverage. its results can be confirmed “visually” with

    tmx -h -v -v

(that’s 2 verbose), which shows a “deep tree” listing that currently
includes only the mounted one-offs but may one day include more
(maybe with more verboses). this mega-listing is :[#tmx-018.B].

we have done it for now without coverage because we use its list
as a guide to drive further development, including coverage of itself..

Times

these times go back almost a week. we assume the days are contiguous
but not sure, we should probably start including those. the value
here is diminished because these aren’t all end-to-end sessions..

  03:53  begin mountable one-offs!!!
  06:34  all the one-offs without help screens
  12:52  start
  13:31  mock page looks good populated with real one-off names and mocked descriptions
  23:11  made massive output files of all of 27 one-off help screens
  23:38  begin three laws on the scraper
  16:13  about to run first test with real scraper
  20:42  first two test cases appear OK - going to write script to run the gamut
  22:42  finished 28 cases semi-visually ("granite")
  03:00  "granite" green and committed
  14:59  “opal” committed

reminder:
  tmx test-all -slice 1st half
  tmx test-all -slice 2nd half

(2 in 34.03)