(-3.175 [*] housecleaning: somewhat unify stream & scanner (.. 4266))
Generally: the `push_stream` and `concat_stream` that we had in a
few places was terrible (explained in [#co-016.2]). this gave us an
itch to scratch in the form of wanting our (what we now call)
“compound stream” and “compound scanner” classes. (very simple!)

but for that we wanted a RISC-like architecture for streams (that
is, for whatever internal medium it is you want to stream (arrays,
filehandles, some event buffer), write an adapter with ONLY ONE
method: `gets`. then include the instance methods and you get one
dozen plus useful map-reduce-expand methods “for free”.

(another itch we wanted to scratch was to abandon using a dedicated
file for our main `Stream` class. we are tying to keep `common.rb`
SLOC “low”, but streams are used so frequently, the jump into
autoloading started to get annoying.)

this drew our attention to the state of things with some *ancient*
code around the earliest manifestations that streaming took (we first
called them “scanners”, before we appreciated a difference between
what we now call streaming vs scanning (the later is the former with
one peek). then at an intermediate phase we went crazy and decided
to call *everything* streams, even what we now call scanners. (we then
called them “polymorphic streams”, a mouthful that endured surprisingly
long).

while we were at it we modernized some other commonly used things that
we saw along the way.

Design Axioms:

  - no more `Split_Personality_Case` for the commonest nodes
    (`Bound_Call` was really bothering us).

  - old nodes that have endured this long can go public

  - old nodes that became obsolete or feature-islands can sunset.

Design Consequences:

  - for streams and scanners

    - renamed `PolymorphicStream` to `Scanner`

    - unify scanner method names with newer argument scanner method
      names somewhat, so no `current_token`, `head_as_is` instead.

  - of singletons:

    - THE_EMPTY_SCANNER because it's too easy not to do it this way
    - THE_EMPTY_STREAM (with some pain) for symmetry

  - for performer and model support,

    - got rid of `Actor` for `Actor::Monadic` etc
      (we eventually want to melt all “actor” for “magnetic”)

    - DRY all `SimpleModel` implementations to 2.
    - abstract `MagneticBySimpleModel` naturally

  - for `Box`
    - “keys” not “names” (“name” suggests human readable, which it
      is not necessarily.) our antipathy against sounding like a
      hash is gone.

    - "offset" not "position", to be consistent with elsewhere
      (no matter what don’t say “index” which usually refers to
      a cached structure holding a collection of metadata.)

  - of formerly private sidesystem sub-sub nodes:
    - promote them. they have stabilized.
      `Distill`, `LineStream_via_*`, `WordWrappers`, `Yamlizer`

Times:

  01:05  start to play around with stream changes
  14:12  next day
  14:53  begin write new test for random access immutable
  02:55  now

(2 in 22.95)