(-3.164 [pl]<={[pl],[ba]} renames.. )
fortunately for us, the bulk of [pl] (and also this piece of [ba] as
well) is ALL "feature islands" (i.e asset code covered by test code
but not used anywhere in "production" (i.e by any "applications")).

the thing about feature islands is they can be repurposed freely
without needing to go thru the pain of "redlisting" and "greenlisting"
effected applications (because there are no applications to break).

generally

  - the (perhaps oldest) legacy plugin "framework" called
    "dependencies" will become the new, simple "lazy index"

  - the "queue" of [ba] was salvaged from [hl] before it was
    sunsetted (probably); that will become the conceptual home
    of a related idea for new plugins.

  - (the "filesystem based [plugin collection]" was kind of a
    bump at the toplevel in light of our upcoming "plugin manifesto"..
    so we have re-housed it as an "operator branch" (it's not) and
    put it in magnetics..)