((-0.885 [ts] starting the quickie service (( 10 )) ))
Discussion: here is where the “philosophy” of the new architecture
comes to the fore. what once was memoized devilishly as variables
in closure or ivars in modules, we now use plain old classes in the
ordinary way for such variables. (for one thing it makes testing
possible, for another it should look more familiar generally).

  - spike mock module with no coverage!

  21:07  begin (reapplying old new code that once worked)
  22:05  tests passing but asset node needs cleanup
  22:31  begin to commit
.