((-0.847 [ts] quickie becomes self-supporting (again) (( +1=11 )) ))
these tests are the first tests in this cluster that can be run with
either rspec or quickie. currently, to run them with quickie you have
to run them one at a time. we keep going back and forth on whether
running multiple files should be part of the quickie “recursive runner”
(next cluster)..

the test that touched “examples” is only one example in one context
with no predicates.

(we had to joggle it around the order a bit so that ARGV wouldn’t be
parsed unless quickie had reign.)

  - [ze] eek not covered niCLI fail-early additions

  21:07  begin reinstating the (once working) new code to run examples
  22:05  now - tests passing but asset node needs cleanup
  22:31  done cleaning..
  04:00  finished rewriting history to bring some legacy pieces fwd
         so we delete them at the right time. now, try self-support..
         (slept)
  15:17  now
.