((-0.539 [ts] quickie “onefile” API!! a wild idea appears.. (( +7=30 )) ))
this was a wild experiment that went unexpectedly successfully.
the idea is the same as always: we want an “API-style” client
instead of a CLI client, because testing with a CLI client can
get ridiculous, and an API client is closer to our platonic ideal
“microservice”.

(more specifically, we want to use an API-style interface to test
all the error permutations and boolean permutations of tags and
line ranges, etc..)

but this is trickier in [ts] quickie because the traditional CLI/API
“invocation” model is halfway out the window because we have the
supposedly long-running runtime which holds the supposedly long-
running service which must be started “around” a client. whew!

anyway, in toughing through developing an API-style interface
we inadvertently improved on something: what was most useful for
testing was not to take a list of files (for the API “operation”)
but to take a single proc that loads all the code with tests. a
fantastic side-effect of this is that you can load files *in* this
proc and it should just work. a fantastic side effect of *this*
is that it scales better than loading all the files at once: each
file that is loaded (for those test files that encompass a single
`describe` block, which is most of them), at the time it reaches
the end of the `describe` block this is when the tests in that
block are run; and so on for each of possibly many many files.
the only thing that needs to remain memory-persistent across all
these test file  invocations is the “statistics aggregation”, which
is trivially small.

now, to be clear, we don’t yet know if rspec is doing the bad thing
and front-loading the test trees (the example nodes and any recursive
contexts nested in every `describe` block) across EVERY file before
it runs one test, but if it is, OOOO NELLLY

  22:51  without any warning, start making an API
  23:30  calling API but doing nothing.
  10:46  API test file alone is passing with flying colors.
         need to re-seat sizeable API work into other file.
  11:34  now.

(we aren’t bothering self-supporting yet for dev because we don’t
 yet have line numbers, which we use most of the time.)

rspec test_support/test/10-quickie
.