(-3.109 [*]            renames for semi-unify expag.. )
Generally the new "design objective" here (just the new plain idea,
really) is that while "expression agent" is a strong idea we are sticking
with, it should not come to stand for a monolithic interface (and by
interface here we mean in the java sense of "set of methods"). rather,
"expression agent" is the abstract idea, and one particular kind of
expression agent is the "interface expression agent". it is this latter
one that confers a particular set of methods.

  - [hu] EN "methods" wants to be this new "simple inflection session"
    idea: generally, keep members and state out of expression agents.
    the "simple inflection session" should be a plain old class.
    this crazy methods enhancer is interesting but it's novelty
    in the wrong direction. it takes steps backwards by cluttering
    the client with more methods, when it would be better to use
    composition and dedicated agent interfces (coming up next)

  - also, whenever such nodes have dedicated files we should just
    go ahead and give them a public name (no trailing underscores)
    because it's such a strong convention that we want the name to
    convey this connectedness to a greater whole.
.