(-18.827 [*] begin “E.K” (.. 4390))
Generally, we need [br] “entity” to go away (too much meta, too fragile,
too hard to debug, too much API to figure out how to extend it) but
we don’t want to do anything that “feels” too drastic on the numerous
(14-ish?) clients that use [br]. (more broadly we want [br]-as-framework
to go away: hard to debug, locked in, should be compatible with [ze],
[ze] should subsume it.) As such:

  - rather than a “framework”, introduce a series of small “toolkits”
    that can eventually repeal & replace [br]-as-framework.

  - we’d like the new facilities to allow the old clients to maintain
    the same ‘look-and-feel’ (as in code), so

  - keep the same ‘Actions’ and model-centric architecture
    (not “operations”, but it’s just semantics)

  - keep the same or a similar attribute grammar.

To this end we will need:

  - a new way (a cleaned up way) to navigate an application tree
    of arbitrary complexity “recursively” (while still letting nodes
    be autonomous enough to introduce their own “operator branches”)

  - a cleaned up way of modeling properties

  - a cleaned up way of effecting normalization against those
    properties.

we ended up doing all of this in this commit by:

  - injecting new “entity killer parameter” into defined attribute

  - along with new “meta-attributes” into old.

these above two were short and sweet, easy to implement.

   - inject new “normalization” into old

this one was harder, and came with a full algorithm.
leading up to that we also cleaned up all old tests for “attribute”
and “meta-attrubte” that were using this twisted sandbox module.

  00:47  cleaned up meta-attribute tests
  04:05  iambic grammar worked, then moved from [fi] to [pl]
         (also spent a while nuberizing tests for [pl])
  12:05  time to wrap it up - the first EK commit
  22:15  now

(2 in 20.43, 20.57, 20.91, 21.44)