(-18.826 [sn]<-[*] begin microservice-ize API (.. 4329))
Synopsis: begin migrating sidesystems off of [br] WEEEE!

Notes about peripheral side-systems affected:

  - [ze] we in effect rewrote “module-centric operator branch”
    because what was there was just a too-hard-to-read sketch.

  - [*] everybody (almost) needs a `subject_API_value_of_failure`
    now that `NIL` is replacing `FALSE` as the value that an API
    results in when it fails..

Details and discussion:

NOTE - there will be a permanent-seeming drop in the number of tests
because we’re postponing the restoration of CLI to much later so that
we can focus on that all at once across several sidesystems.

we’re breaking this leg up into two steps because of its potentially
large scope what with this being the first commit of our attempts at
weening off of legacy [br] and into using our microservice toolkit(s)..

we don’t really have anything interesting to say about the work itself,
other than that it’s satisfying to bring the new stuff in.

regarding the craziness that’s in “.stack” right now:

first off, here’s how we got the list of files:

    q -list-files -order n-1 snag/test

(“q” is “test_support/bin/tmx-test-support-quickie”)

(reminder: “n-1” is the way we can get the listing in reverse
(“N down to 1”) so that the easiest tests are last, i.e show
up in the frontmost in the stack.

then we went though one-by-one (somewhat) at tried to fix most
“models” tests as best we could. (we had to skip the
“criteria integration” test file, for example, because that involves
a lot of other “silos”.)  files that were areen we marked with
“OK” in the “first column” of the stack file, and the others are
marked with some string other than “OK”.

then, with the current “.stack” file, here’s the most amazing
one-off I have yet done:

    q -wip-files $( tail -n32 -r .stack | awk 'BEGIN{ORS=" "} {if("OK"!=$1){print $2}}' )

(the ugly part is the “32” - that’s the number of relevant trailing
lines in the stack file, which is the number of test files in [sn].)

this says, using an output record separator (“ORS”) of space,
print every second column (“cel”) of every line whose first column
(“cel”) is not the string “OK”.

this list of files (as one line string) is then the argument to
“quickie -wip-files”.

    tmx test-all -order after -reverse -slice 1st half
    tmx test-all -order after -reverse -slice 2st half

(2 in 24.57)