(-18.497 [*] rename all “load ticket (etc)” to “reference” (.. 4208))
Discussion: the one thing in common that
“byte {up|down}stream identifiers” and node tickets and load tickets
have is that they are not the thing but a reference to the thing; and
they can all be de-referenced. so this years and years long aphasia
(remember “sheet” as in “character sheet”?) can finally settle on a
consistent, self-documenting name. (the “ticket” metaphor was a
perfectly good one, but it was a bit more of an idiom than it needed
to be.)

Also:

  - “byte stream” (upstream and downstream reference) abstracted
      from [br] and put into [ba].

(2 in 16.33, 16.04)