(-18.425 [fi] refactor to nudge towards value normalizer injection (.. 4202))
Gripes and Musings:

in a branch that we will probably destroy now, we began an effort to
make [br]-era models and entities use the “one ring” algorithm. halfway
(and a half day) into this effort we discovered how entrenched the old
way is. to force this old stack to use our new “provisions” would entail
non-trivial rearchitecting of [br], something that makes us ask
“why bother” because we have just embarked on the effort to ramp out of
it.

as a perfect compromise, we now have only one place where we traverse
over associations of the “remaining extroverted stream”. (this same
method is used whether we are doing an argument-driven parse or
“normalize-in-place”.) this one place uses as a sort of “injection”
a simple proc that takes one argument, the association.

this proc represents whatever “value normalizer” we resolved given
the state of things. this proc is normally a bound method from
“one ring” (we considered breaking out the value-normalizing parts
into its own injectable class, but decided against it for now because
of the appeal and convenience of having the n11n happen mostly in
one object. (our current threshold is you cannot have more than two
`@_current_*`-type ivars.)

but this arrangement does allow for us to inject the legacy
“value normalizer” algorithm, one that as we have discovered is rife
with provisions that contrary to the ideal “one ring” normalization.

(2 in 19.52)