(-18.424 [fi] refactor to “association source” (.. 4202))
Design objective: there was too much spaghetti coming from this
“method reference” technique.

Design consequence: injection of how association streams are
represented and normalized against.

Discussion: this idea sprung up from when we tried to assimilate
“facility C”: abstract out of the “one ring” algorithm itself (that
is, make “injectable”) how a stream of associations is produced
**and how they are normalized**. because there’s deeply entrenched,
“old school” algorithms that use the “entity” as an *unsanitized*
value store, we have to pull out the modern one-ring field
“normalization pipeline” into its own separate guy that is injected
as appropriate.

we don’t get as far as actually assimilating facility “C” here because
it was too much to take on in one commit (it amounted to almost a full
write of the “normalization” node) and things were failing in hard to
understand ways..

(2 in 20.41)