(-18.418 [fi]<-[*] name unification & general cleanup (.. 4201))
as discussed in near previous commits, “attributes” is now just one
application of our “association” “toolkit” that [fi] is becoming.
to emphasize this we generally eschew the use of the term “attribute”
except when that specific sub-library is being used. also we more
generally are trying to give nodes more appropriate names with our
new worldview of how everything fits into “association theory”. so:

  - `MetaAttributes` -> `CommonMetaAssocations` because they aren’t
    the only meta-associations, just a set of implementations for
    common ones.

promote up & out of “attributes” for being thematically unrelated:

  - “struct”
  - “normalization”

general cleanup

  - re-order how the common meta-associations are ordered in the
    file not to be lexical but thematic

(2 in 19.38)