(-18.409 [*] (4/4) magnetize everything (.. 4233))
WHEW!!!!!

Narrative introduction: as we were humming along refactoring [tm] to
wean it off of [br], we wanted to re-use the *mechanism* of [br]
workspace stuff (creating “workspace” config filed and so on) without
needing to do the absolutely ghastly trick of subclassing the action
classes of [br].

the now obvious solution to this is to use the performers of [br]
(then “actors”, now “magnetics”) to do the bulk of the work; the rest
that we should have to add in the new application is just “glue”.
this magnetic/action dichotomy is a perfect fit for this use-case,
because the higher-level UI concerns are still left to the application
to decide, whereas the lower-level implementation logic is shared.

so this lead us to want to publicize those [then actors now magnetics]
of [br] involved in not only this action but other actions whose
implementations we would want to re-use in a manner other than the
ghastly aforementioned subclassing.

*this* then lead us to want to A) flush all pending renames but then
B) while we’re at it, maybe we should instill this kind of change
*everywhere*. that is, re-conceive of what we used to house as
“sessions” and “actors” as “magnetics”, so that there is a universally
consistent and best-yet practice utilized for “performers”. SO:

Description of problem: it became too jarring switching mentally between
newer code that uses magnetics and “magnetic by simple model” and the
older code that uses actors and sessions.

“magnetics” are not a replacement for these previous two patterns, but
rather a refinement on them. it’s not that these other two didn’t work,
it’s just that we feel “magnetics” works better.

  - actors always benefit from and never hurt from being re-named to
    a “magnetic convention” name. all benefit, no cost. (the old-
    school pattern here was to name actors after verbs, but rarely
    (never) is a sole verb descriptive enough to give an adequate
    sense at a glance what that performer does.)

  - most sessions would do well to adopt the “magnetic by simple model”
    interface; and maybe 85% of the time it’s mostly straightforward
    to refactor them into this upgrade.

Other refactoring work along the way:

  - got rid of methods that access modules - this is a legacy pattern
    that we adopted to help allay the pain of renaming- /re-housing-
    oriented refactors, but now we feel it is better to have public
    nodes be plain old public nodes instead of using methods to access
    them because it’s less indirection and more self-documenting.

  - [fi], [hu] scratch an itch we had in the latter that depended on
    the former - we wanted to be able to see the method-based
    associations in an “attributes-actor”-like magnetic without
    having to invade the whole namespace (interface) of it. this is
    now possible because we have abstracted the mechanics of this out
    of the attributes actor methods and put it into its own performer.

  - [hu] make excellent use of an “operator branch” in refactoring
    our index of magnetics (files in a folder).

Deletes along the way:

  - [pa] “function/nonterminal/core.rb” was a would-be smalls node,
    but it proved to be anemic so we moved its only asset to the one
    file where it was used.

Significant renames:

  - the “input adapters/output adapters” pattern in some cases is
    seen as being superseded by the magnetics pattern. (a set of
    either input or output adapters as desired can perhaps be achieved
    as a filter on top of a set of magnetics.)

  - finally put something in its rightful place in [ba] for “struct” -
    move this node in [co]  to there that is used for making struct-like
    classes

(2 in 16.5, 16.63, 17.13, 16.62)